--- a/WFInfo/MainWindow.xaml.cs
+++ b/WFInfo/MainWindow.xaml.cs
@@ -69,6 +69,10 @@ namespace WFInfo
                     Settings.settingsObj["Auto"] = false;
                 Settings.auto = (bool)Settings.settingsObj.GetValue("Auto");
 
+                if (!Settings.settingsObj.TryGetValue("CuttingEdge", out _))
+                    Settings.settingsObj["CuttingEdge"] = false;
+                Settings.autoScaling = (bool)Settings.settingsObj.GetValue("CuttingEdge");
+
                 if (!Settings.settingsObj.TryGetValue("AutoDelay", out _))
                     Settings.settingsObj["AutoDelay"] = 250L;
                 Settings.autoDelay = (long)Settings.settingsObj.GetValue("AutoDelay");