Developed a mode that detects scaling at very high accuracy with little performance draw
1080p sees 50ms to 200ms extra latency
4K sees 100ms to 300ms

Currently disabled unless "CuttingEdge" is enabled in settings